+++ LuccaDevises/ApplicationService.Test/RateCalculatorTest.cs
@@ -14,9 +14,9 @@ namespace ApplicationService.Test
         public void ShouldGet59033yens_WhenGiven550Euros()
         {
             List<Change> listeChanges = new List<Change>();
-            listeChanges.Add(new Change("EUR", "CHF", (decimal)1.2053));
-            listeChanges.Add(new Change("CHF", "AUD", (decimal)1.0351));
-            listeChanges.Add(new Change("AUD", "JPY", (decimal)86.0305));
+            listeChanges.Add(new Change("EUR", "CHF", 1.2053m));
+            listeChanges.Add(new Change("CHF", "AUD", 1.0351m));
+            listeChanges.Add(new Change("AUD", "JPY", 86.0305m));
             var mock = new Mock<IPathCalculator>();
             mock.Setup(pc => pc.Rates())
                 .Returns(listeChanges);
